+++ packages/slate/src/transforms/node.ts
@@ -228,7 +228,9 @@ export const NodeTransforms: NodeTransforms = {
         const path = parentPath.concat(index)
         index++
         editor.apply({ type: 'insert_node', path, node })
+        at = Path.next(at)
       }
+      at = Path.previous(at)
 
       if (select) {
         const point = Editor.end(editor, at)
@@ -422,9 +424,12 @@ export const NodeTransforms: NodeTransforms = {
       // of merging the two. This is a common rich text editor behavior to
       // prevent losing formatting when deleting entire nodes when you have a
       // hanging selection.
+      // if prevNode is first child in parent,don't remove it.
       if (
         (Element.isElement(prevNode) && Editor.isEmpty(editor, prevNode)) ||
-        (Text.isText(prevNode) && prevNode.text === '')
+        (Text.isText(prevNode) &&
+          prevNode.text === '' &&
+          prevPath[prevPath.length - 1] !== 0)
       ) {
         Transforms.removeNodes(editor, { at: prevPath, voids })
       } else {
@@ -589,6 +594,14 @@ export const NodeTransforms: NodeTransforms = {
       }
 
       if (split && Range.isRange(at)) {
+        if (
+          Range.isCollapsed(at) &&
+          Editor.leaf(editor, at.anchor)[0].text.length > 0
+        ) {
+          // If the range is collapsed in a non-empty node and 'split' is true, there's nothing to
+          // set that won't get normalized away
+          return
+        }
         const rangeRef = Editor.rangeRef(editor, at, { affinity: 'inward' })
         const [start, end] = Range.edges(at)
         const splitMode = mode === 'lowest' ? 'lowest' : 'highest'
@@ -629,19 +642,23 @@ export const NodeTransforms: NodeTransforms = {
           continue
         }
 
+        let hasChanges = false
+
         for (const k in props) {
           if (k === 'children' || k === 'text') {
             continue
           }
 
           if (props[k] !== node[k]) {
-            // Omit new properties from the old property list rather than set them to undefined
+            hasChanges = true
+            // Omit new properties from the old properties list
             if (node.hasOwnProperty(k)) properties[k] = node[k]
-            newProperties[k] = props[k]
+            // Omit properties that have been removed from the new properties list
+            if (props[k] != null) newProperties[k] = props[k]
           }
         }
 
-        if (Object.keys(newProperties).length !== 0) {
+        if (hasChanges) {
           editor.apply({
             type: 'set_node',
             path,
@@ -841,7 +858,13 @@ export const NodeTransforms: NodeTransforms = {
 
       const rangeRef = Range.isRange(at) ? Editor.rangeRef(editor, at) : null
       const matches = Editor.nodes(editor, { at, match, mode, voids })
-      const pathRefs = Array.from(matches, ([, p]) => Editor.pathRef(editor, p))
+      const pathRefs = Array.from(
+        matches,
+        ([, p]) => Editor.pathRef(editor, p)
+        // unwrapNode will call liftNode which does not support splitting the node when nested.
+        // If we do not reverse the order and call it from top to the bottom, it will remove all blocks
+        // that wrap target node. So we reverse the order.
+      ).reverse()
 
       for (const pathRef of pathRefs) {
         const path = pathRef.unref()!
@@ -942,6 +965,12 @@ export const NodeTransforms: NodeTransforms = {
           const last = matches[matches.length - 1]
           const [, firstPath] = first
           const [, lastPath] = last
+
+          if (firstPath.length === 0 && lastPath.length === 0) {
+            // if there's no matching parent - usually means the node is an editor - don't do anything
+            continue
+          }
+
           const commonPath = Path.equals(firstPath, lastPath)
             ? Path.parent(firstPath)
             : Path.common(firstPath, lastPath)