+++ packages/slate/test/operations/set_node/remove-null.tsx
@@ -5,17 +5,18 @@ import { Transforms, Editor } from 'slate'
 export const input = (
   <editor>
     <element>
-      <text key />
+      <text someKey />
     </element>
   </editor>
 )
 
+// this is supported for backwards compatibility only; newProperties should omit removed values.
 export const operations = [
   {
     type: 'set_node',
     path: [0, 0],
-    properties: { key: true },
-    newProperties: { key: null },
+    properties: { someKey: true },
+    newProperties: { someKey: null },
   },
 ]